pkg/op-brains/pyproject.toml
[tool.poetry]
name = "op-brains"
version = "0.1.0"
description = ""
authors = ["José Ribeiro <jose@bleu.studio>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
langchain-openai = "^0.1.14"
langchain-text-splitters = "^0.2.2"
langchain-community = "^0.2.6"
weave = "^0.50.9"
faiss-cpu = "^1.8.0.post1"
flask = "^3.0.3"
click = "^8.1.7"
flask-cors = "^4.0.1"
gunicorn = "^22.0.0"
flask-limiter = "^3.7.0"
langchain-anthropic = "^0.1.19"
posthog = "^3.5.0"
python-dotenv = "^1.0.1"
op-artifacts = { path = "../op-artifacts", develop = true }
op-core = { path = "../op-core", develop = true }
tqdm = "^4.66.5"
langchain-voyageai = "^0.1.1"
langchain-core = "~0.2"
aiocache = "^0.12.3"
google-generativeai = "^0.8.0"
json-repair = "^0.24"
sentence-transformers = "^5.2.3"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
pytest = "^8.2.2"
pytest-mock = "^3.14.0"
pyright = "^1.1.375"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
